+++ app/repositories/models/__init__.py
@@ -1,1 +1,6 @@
-from .models import *
+from app.plugins import db
+from .beverage import Beverage
+from .details import IngredientOrderDetail, BeverageOrderDetail
+from .ingredient import Ingredient
+from .order import Order
+from .size import Size
